Created single authorization controller for instant action
