+++ Example/Authenticator/View Models/Authorizations/AuthorizationsViewModel.swift
@@ -42,12 +42,19 @@ enum AuthorizationsViewModelState: Equatable {
     }
 }
 
+protocol AuthorizationsViewModelEventsDelegate: class {
+    func reloadData()
+    func shouldDismiss()
+    func scroll(to index: Int)
+}
+
 class AuthorizationsViewModel {
     private struct Images {
         static let noAuthorizations: UIImage = UIImage(named: "noAuthorizations", in: .authenticator_main, compatibleWith: nil)!
         static let noConnections: UIImage = UIImage(named: "noConnections", in: .authenticator_main, compatibleWith: nil)!
     }
 
+    weak var delegate: AuthorizationsViewModelEventsDelegate?
     var state = Observable<AuthorizationsViewModelState>(.normal)
 
     var dataSource: AuthorizationsDataSource!
@@ -55,18 +62,13 @@ class AuthorizationsViewModel {
 
     private var poller: SEPoller?
     private var connections = ConnectionsCollector.activeConnections
-    var authorizationFromPush: (connectionId: String, authorizationId: String)? {
+
+    var singleAuthorization: (connectionId: String, authorizationId: String)? {
+        willSet {
+            setupPolling()
+        }
         didSet {
-            if let detailViewModel = dataSource.viewModel(
-                by: authorizationFromPush?.connectionId,
-                authorizationId: authorizationFromPush?.authorizationId
-            ),
-            let index = dataSource.index(of: detailViewModel) {
-                state.value = .scrollTo(index)
-            } else {
-                state.value = .presentFail(l10n(.authorizationNotFound))
-            }
-            authorizationFromPush = nil
+            getEncryptedAuthorizationsIfAvailable()
         }
     }
 
@@ -111,10 +113,16 @@ class AuthorizationsViewModel {
             success: {
                 detailViewModel.state.value = .success
                 detailViewModel.actionTime = Date()
+                after(3.0) {
+                    self.delegate?.shouldDismiss()
+                }
             },
             failure: { _ in
                 detailViewModel.state.value = .undefined
                 detailViewModel.actionTime = Date()
+                after(3.0) {
+                    self.delegate?.shouldDismiss()
+                }
             }
         )
     }
@@ -130,30 +138,38 @@ class AuthorizationsViewModel {
             success: {
                 detailViewModel.state.value = .denied
                 detailViewModel.actionTime = Date()
+                after(3.0) {
+                    self.delegate?.shouldDismiss()
+                }
             },
             failure: { _ in
                 detailViewModel.state.value = .undefined
                 detailViewModel.actionTime = Date()
+                after(3.0) {
+                    self.delegate?.shouldDismiss()
+                }
             }
         )
     }
 
     private func updateDataSource(with authorizations: [SEAuthorizationData]) {
         if dataSource.update(with: authorizations) {
+            delegate?.reloadData()
             state.value = .reloadData
         }
 
-        if let authorizationToScroll = authorizationFromPush {
+        if let authorizationToScroll = singleAuthorization {
             if let detailViewModel = dataSource.viewModel(
                 by: authorizationToScroll.connectionId,
                 authorizationId: authorizationToScroll.authorizationId
             ),
             let index = dataSource.index(of: detailViewModel) {
+                delegate?.scroll(to: index)
                 state.value = .scrollTo(index)
             } else {
                 state.value = .presentFail(l10n(.authorizationNotFound))
             }
-            authorizationFromPush = nil
+            singleAuthorization = nil
         }
     }
 }
@@ -161,9 +177,11 @@ class AuthorizationsViewModel {
 // MARK: - Polling
 extension AuthorizationsViewModel {
     func setupPolling() {
-        poller = SEPoller(targetClass: self, selector: #selector(getEncryptedAuthorizationsIfAvailable))
-        getEncryptedAuthorizationsIfAvailable()
-        poller?.startPolling()
+        if poller == nil {
+            poller = SEPoller(targetClass: self, selector: #selector(getEncryptedAuthorizationsIfAvailable))
+            getEncryptedAuthorizationsIfAvailable()
+            poller?.startPolling()
+        }
     }
 
     func stopPolling() {
